Refactor graph helpers from Svg.Database to Models.Graph
